Updated input model
Input model now uses joystick/controller input, 2 input vectors, and
outputs 1 error vector.

The first input vector is for constants to multiply the controller axes
by to obtain degrees.  This may also be considered a limit for the
input.  For example, if the roll constant is 45 degrees the user cannot
exceed that on purpose.